lru: use weak pointers in LRU stack
Galaxycache isn't always the most friendly to the Go Garbage Collector.
By keeping a whole pile of pointers in a linked list that the GC needs
to scan in its mark phase, we're compromising the tail latency of any
processes that maintain large numbers of objects in their caches.

Fortunately, Go 1.24 provides weak pointers, so as long as we keep an
"owning" reference in the cache's map, we can make all next/previous
pointers weak pointers and avoid a whole bunch of extra GC work in its
mark phase. (note: this optimization requires that the weak pointers be
at the end of the struct so the GC can skip scanning them if there are
pointers in the value -- if there aren't, then the GC can skip the
linked list entries entirely -- not likely unless one has integer keys)