jsoncodec: add a codec impl using encoding/json
While protobuf is almost always a better choice for payload marshaling,
there are cases where galaxycache is being added at short-notice to
systems that already have canonical JSON representations. In that case,
defining a protobuf message-type for each Galaxy would delay the
benefits of having Galaxycache in-place.

Note: this is mostly a C&P of the protocodec's V2 types with a few
adjustments because we can be a bit less picky here.