k8swatch: SetIncludeSelf management
Add a couple methods to the universe to support and add code to k8swatch
that handle updates to the self pod-name.

Due to the slight weirdness of changes that cross modules, we use a
type-assert and progressive enhancement.

After we cut a Galaxycache release with the new methods, we can update
go.mod in the k8swatch submodule.