lru: rewrite FuzzTypedAddRemove
Vastly simplify FuzzTypedAddRemove by limiting the keyspace, and make it
more useful by providing triggers for running the GC in parallel.

Use the two high-bits to select the operation and the low-6 bits the
key. (this eliminates the degree of freedom where the fuzzer picks
super-long keys, since it's effectively fuzzing the hash-table's
hash-function.)

Unfortunately, this does cause the fuzzer to explore the control flow in
the Garbage Collector somewhat, (that's an unfortunate random walk),
but when working with weak pointers, you really do need the GC to run to
have any confidence that you didn't remove something you need.