consistenthash: upper-bound ownership: fix GetReplicated
Document that we're storing _*upper*_ bounds of each segment and switch
prevSegmentOwner to nextSegmentOwner to make the hashring ownership
traversal consistent.

The `prevSegmentOwner` method was intended to find the owner of the
object if the actual owner of that segment was excluded. Unfortunately,
since the segment ownership wasn't documented, when writing that method,
we misunderstood the behavior, and ended up implementing it assuming
that the keys were lower-bounds rather than upper-bounds.

Switch this to handle upper-bounds (traverse forward by a slot) instead
of lower-bounds (traversing backward by a slot).

This is (questionably) non-impacting for the only known user of the
GetReplicated method, but that has enough redundancy that it should be
fine while upgrading -- it only matters when later replicas hash to
segments owned by already selected owners and we need to iterate to find
another owner for that replica.

Note: fixing this makes implementing a datastructure for efficiently
implementing ownership tests for a specific owner significantly easier.
(It also makes the implementation conceptually consistent.)