PeerPicker: add peek picking
Add a new hashring to the PeerPicker that always excludes "self".
Additionaly, add a method so we can pick the peer that should receive
a specific Peek call.

Make some adjustments to the tests and plumb a clock in through the
universe.

Also, bump go.mod to go 1.23 so we can use chtest in the root package's
tests.