chtest: move in-test check to its own file
The backwards-compatiblity tests we'll be adding will be much easier to
work with if they can use the chtest package, but since they'll be
separate modules, it'll be useful to have the chtest package.

Since those will be run with `go run` via os/exec, we won't have any
problem setting `-tags`, so we can set a nice, long. scary-sounding
build-tag to gain access to chtest outside a test.