grpc: Peek method
Add a `PeekPeer` gRPC method that provides the Peek functionality.

For uniformity, switch to calling the new `GetWithOptions` method. (it
doesn't do anything different for the moment, because only FetchModePeek
is implmented, but the next commit will implement Peek in the Galaxy, at
which point this will prevent distributed deadlocks if anything is
mismatched in the hashring.

The tests have not been updated yet (beyond fixing some staticcheck nits
about deprecated grpc calls).