chtest: fix segments_per_key calculation
When writing the initial version of this package, I did a back of the
envelope calculation, and forgot to draw myself a full diagram for the
n=2 and n=3 cases, meaning that I didn't actually _count_ how many
entries I needed with the layout I had. Writing it down and counting
showed that I was off by a factor of 2. (which also jives with the sizes
of the internal maps/slices this package creates (which are
unfortunately only visible in a debugger because they're only propagated
by the hash-function closure's capture)