compatibility tests: gRPC & HTTP tests
Add compatibiility tests that use multiple versions of Galaxycache
(currently v1.2 and the current commit). It also includes peeking, mixed
instance setups and both HTTP & gRPC fetchers.)

In particular, it uses a submodule to provide a test binary that's
forced to use an old version of galaxycache. (unless go.work forces our
hand in some tests)

This approach can be expanded to other versions as we continue to update
the protocol.