lru: add Expiry support for go 1.24+
Add expiry support for go 1.24+ with the help of build-tags and weak
pointers.

Several key type-definitions and methods are now duplicated between a
pre-1.24 and 1.24+ file, including Add, Get, and the TypedCache type
itself.

Care has to be taken to make sure that Get doesn't return an expired
value (and Add..., Remove... and friends are also configured for removing any
expired entries before removing anything from the LRU stack).