Galaxycache: new BackendGetterWithInfo iface
Add a new BackendGetterWithInfo interface that allows a BackendGetter
implementation to return an expiration time as well as functioning with
the normal Codec mechanism for extracting the values.

Plumb the BackendGetInfo with a new method on Universe, additional
return values on Fetch & Peek, and update tests+http&grpc impls.

Follow-up commits will update the grpc and http methods to propagate
expirations.