cachekey: remove ID length assumption
AppendStrID was copied out of a system that only used UUIDs for string
IDs, so we could make the assumption that IDs would be shorter than
127-bytes. This is not true generally, but apparently while copying the
function, we forgot to re-evaluate the assumptions in the code. (even
when stated in the comments)

Fortunately, `encoding/binary.AppendUvarint` will grow the buffer if
necessary, so this is a performance bug, not a correctness bug.